INF-1400
git-svn-id: https://ikanow.jira.com/svn/INF/branches/infinit.e-published-to-github@5673 e6edf6fb-f266-4316-afb4-e53d95876a76

+++ deploy/rpm/infinit.e-processing-engine.rpm/SOURCES/infinit.e-processing-engine/mnt/opt/infinite-home/bin/infinite-px-engine.sh
@@ -4,19 +4,24 @@ LAST_CLEANSE="0"
 RESET_FILE="/opt/infinite-home/bin/RESET_FILE"
 SYNC_FILE="/opt/infinite-home/bin/SYNC_FILE"
 STOP_FILE="/opt/infinite-home/bin/STOPFILE"
+ALLSTOP_FILE="/opt/infinite-home/bin/ALLSTOPFILE"
 EXTRA_JAVA_ARGS=" -Xms2048m -Xmx2048m -Xmn512m -Dcom.sun.management.jmxremote -Dsun.net.client.defaultConnectTimeout=30000 -Dsun.net.client.defaultReadTimeout=30000 -classpath infinit.e.harvest.library.jar:*:infinit.e.core.server.jar"
+
+if [ -f $ALLSTOP_FILE ] && [ -z "$1" ]; then
+	exit 0
+fi
 
 harvest()
 {
-	echo "Starting harvest"
+	echo "Starting harvest $1 $2"
 	#do harvest cycle
 	if [ -f "$RESET_FILE" ]; then
 		echo "Reseting Harvester"
-		java ${EXTRA_JAVA_ARGS} com.ikanow.infinit.e.core.CoreMain --harvest --config $CONFIG_LOCATION --reset
+		java ${EXTRA_JAVA_ARGS} com.ikanow.infinit.e.core.CoreMain --harvest --config $CONFIG_LOCATION --reset $1 $2
 		rm -f $RESET_FILE
 	else
 		echo "Doing normal harvest"
-		java ${EXTRA_JAVA_ARGS} com.ikanow.infinit.e.core.CoreMain --harvest --config $CONFIG_LOCATION 
+		java ${EXTRA_JAVA_ARGS} com.ikanow.infinit.e.core.CoreMain --harvest --config $CONFIG_LOCATION $1 $2
 	fi
 }
 
@@ -40,14 +45,14 @@ cd $LIB
 # Run the harvester
 LAST_CLEANSE="$(date --date '7 days ago' +%s)"
 while true; do
-	harvest
-	if [ -f "$STOP_FILE" ]; then
+	harvest $1 $2
+	if [ -f "$STOP_FILE" ] || [ -f "$ALLSTOP_FILE" ]; then
 		exit 0
 	fi
 	if [ -f "$SYNC_FILE" ]; then
 		optimize
 		sync
-		if [ -f "$STOP_FILE" ]; then
+		if [ -f "$STOP_FILE" ] || [ -f "$ALLSTOP_FILE" ]; then
 			exit 0
 		fi
 	fi
